requirements.txt
fastapi
python-multipart
uvicorn[standard]
motor
pymongo[motor]
pypdf
pydantic
python-dotenv
pydantic-settings
celery[redis]
redis
langchain
# langgraph-core is not on PyPI; install via GitHub when available.
 # e.g.: pip install git+https://github.com/langgraph/langgraph-core.git
weasyprint
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.5
httpx==0.25.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
